COMP: fixed some 64 to 32 bit conversion warnings

+++ VTK/Rendering/vtkOpenGLActor.cxx
@@ -31,7 +31,7 @@ vtkStandardNewMacro(vtkOpenGLActor);
 // Actual actor render method.
 void vtkOpenGLActor::Render(vtkRenderer *ren, vtkMapper *mapper)
 {
-  float opacity;
+  double opacity;
 
   // get opacity
   opacity = this->GetProperty()->GetOpacity();
